add per-site user styles
since the css domain selectors do not work in webkit, another approach needs to
be taken to support custom css for specific sites. In config.h there is now a
list of regexes and corresponding css filenames. When a page is loaded and
userstyles are active, the first entry that matches the uri is used as
user-stylesheet-uri. The old static stylesheet is removed but can still be used
by installing a default rule matching any site at the end of the list.